kubeflow: v2 e2e training pipeline
Remove the requirement to upload a config.mk to s3, as that is
only used for the Thingpedia developer key and we can hardcode that.

Ensure that all the owner, project, bucket, etc. options as
passed to make, and remove the defaults from the pipeline so
users are required to set them when creating a run.

Update sync-repos for future compatibility with npm instead of yarn.

Remove eval version from evaluation.